+++ src/Bridge/Gemini/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 * [BC BREAK] `ResultConverter` now returns a `MultiPartResult` when there are multiple `parts` in a `candidate`
 * [BC BREAK] `ResultConverter` now `ExecutableCodeResult` and `CodeExecutionResult` parts when using `code_execution` server tool
 * [BC BREAK] Throwing when code execution server tool fails is replaced with `CodeExecutionResult::isSucceeded()`
+* Add possibility to pass `tool_config` to the model
 
 0.7
 ---